reactor: 解耦， 分割功能
