+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wedecode",
-  "version": "0.3.1",
+  "version": "0.5.0",
   "type": "module",
   "description": "微信小程序源代码还原工具, 线上代码安全审计",
   "bin": {
@@ -15,7 +15,7 @@
     "unlink": "pnpm unlink",
     "release:npm": "vite build && npm publish",
     "release:git": "vite build && git commit -am v$npm_package_version && git tag $npm_package_version && git push --tags ",
-    "test: cmd": "wedecode",
+    "test:cmd": "wedecode",
     "dev:unpack:file": "DEV=true wedecode --overwrite=true -o OUTPUT pkg/mini-game/_subpackages_common_.wxapkg",
     "dev:unpack:file:nooverwrite": "DEV=true wedecode -o OUTPUT pkg/issues2/_pagesA_.wxapkg",
     "dev:unpack:dir": "DEV=true wedecode -ow true -o OUTPUT pkg/game-wxb98ac240fd74b0e3",
